requirements.txt
fastapi
uvicorn
chromadb>=0.4.0
openai
pydantic
python-dotenv
pdfplumber
psycopg2-binary
redis
celery
requests
sqlalchemy
sentence-transformers
pydantic-settings
